Created Initial endpoints for help, super, mls

+++ src/main/java/ausfin/api/records/TaxResult.java
@@ -1,4 +1,4 @@
 package ausfin.api.records;
 
-public record TaxResult(Integer baseTax, Integer taxRate, Integer totalTax, Integer afterTax) {
+public record TaxResult(Integer baseTax, Float taxRate, Integer totalTax, Integer afterTax) {
 }
